[feature] handler PUBLISHED -  save_citation no more cleery task
