+++ jdhapi/tasks.py
@@ -8,7 +8,6 @@ from .models import Article
 from jdhapi.utils.articles import (
     get_notebook_stats,
     get_notebook_specifics_tags,
-    get_citation,
     generate_tags,
     generate_narrative_tags,
     get_notebook_references_fulltext,
@@ -72,18 +71,6 @@ def save_article_specific_content(article_id):
         logger.error(f"save_article_specific_content:{article_id} not found")
     data = get_notebook_specifics_tags(article, raw_url=article.notebook_ipython_url)
     article.data = data
-    article.save()
-
-
-@shared_task
-def save_citation(article_id):
-    logger.info(f"save_article_citation:{article_id}")
-    try:
-        article = Article.objects.get(pk=article_id)
-    except Article.DoesNotExist:
-        logger.error(f"save_article_citation:{article_id} not found")
-    citation = get_citation(raw_url=article.notebook_ipython_url, article=article)
-    article.citation = citation
     article.save()
