+++ jdhapi/views/articles/articles.py
@@ -4,7 +4,7 @@ from django_filters.rest_framework import DjangoFilterBackend
 from rest_framework import generics, filters
 from rest_framework.permissions import BasePermission
 from django.shortcuts import get_object_or_404
-from jdhapi.views.articles.status_handlers import TechnicalReviewHandler, CopyEditingHandler, PeerReviewHandler
+from jdhapi.views.articles.status_handlers import *
 from rest_framework.permissions import IsAdminUser
 from rest_framework.response import Response
 from rest_framework.views import APIView
@@ -96,6 +96,7 @@ class ArticleStatus(APIView):
         'TECHNICAL_REVIEW': TechnicalReviewHandler(),
         'COPY_EDITING': CopyEditingHandler(),
         'PEER_REVIEW': PeerReviewHandler(),
+        'PUBLISHED' : PublishedHandler()
     }
     
     def patch(self, request, abstract__pid):
